requirements.txt
streamlit
supabase
pytesseract
cryptography
python-dotenv
Pillow
opencv-python-headless
qrcode[pil]
fpdf2
altair
pytest
